// features/auth/register/model/schema.ts
import { z } from "zod";

export const formSchema = z.object({
  login: z
    .string()
    .min(4, "Минимальная длинна логина 4 символа")
    .max(16, "Максимальная длинна логина 16 символов"),
  email: z.string().email("Введите корректную почту"),
  password: z.string().min(6, "Минимальная длинна пароля 6 символов"),
  name: z.string(),
});